@@ -15,7 +15,8 @@
 from __future__ import annotations
 
 import math
-from typing import Any, Callable
+from typing import Any
+from collections.abc import Callable
 
 import boxes
 from boxes import Boxes, edges
