@@ -1,7 +1,8 @@
 from __future__ import annotations
 
 from math import *
-from typing import Any, Callable
+from typing import Any
+from collections.abc import Callable
 
 from boxes import vectors
 
